Define a function print_triangle that takes two parameters:
When called, the function should output a centred triangle shape made up of the
desired symbol. The number of symbols in each row should increase by two
with each level, starting with one symbol on the first level, then three on the
second level and so on until the final level is reached. For example, for
number_of_levels = 4 and symbol = ”*” the following triangle should be
produced: